+++ RELEASE-NOTES.md
@@ -60,7 +60,7 @@
 
 ## v2018.08.5
 
-- Fingerprint Card Identification: set samis id to idNumber in case of personInfo is null.
+- Fingerprint Card Identification: set samis id to documentId in case of personInfo is null.
 
 ## v2018.08.4